requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
pydantic
joblib
scikit-learn
python-dotenv
